Factor out getNodesOfKind
